Fixed bug about converting types during training

+++ dataset.py
@@ -161,8 +161,9 @@ class MXFaceDataset(Dataset):
         header, img = mx.recordio.unpack(s)
         label = header.label
         if not isinstance(label, numbers.Number):
-            label = label[0]
-        label = torch.tensor(label.astype(np.int32), dtype=torch.long)
+            label = label[0]        
+        # label = torch.tensor(label.astype(np.int32), dtype=torch.long)
+        label = torch.tensor(label, dtype=torch.long)
         sample = mx.image.imdecode(img).asnumpy()
         if self.transform is not None:
             sample = self.transform(sample)
